refactor(e2e tests): merge single with multi-instance tests
