+++ tests/__data__/docusaurus2-graphql-doc-generator-groups.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  id: "schema_with_grouping",
   schema: "data/schema_with_grouping.graphql",
   rootPath: "./docs",
   baseURL: "group-by",
