+++ tests/__data__/docusaurus2-graphql-doc-generator-multi-instance.config.js
@@ -1,0 +1,4 @@
+module.exports = [
+  require("./docusaurus2-graphql-doc-generator.config.js"),
+  require("./docusaurus2-graphql-doc-generator-groups.config.js"),
+];
