+++ tests/__data__/docusaurus2-graphql-doc-generator.config.js
