--- tests/e2e/docusaurus2-graphql-doc-generator-multi-instance-1.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  id: "tweet",
-  schema: "data/tweet.graphql",
-  rootPath: "./docs-tweet",
-  baseURL: "schema",
-  linkRoot: "/",
-};
