--- tests/e2e/docusaurus2-graphql-doc-generator-multi-instance-2.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  id: "swapi",
-  schema: "data/swapi.graphql",
-  rootPath: "./docs-swapi",
-  baseURL: "schema",
-  linkRoot: "/",
-};
