--- tests/e2e/docusaurus2-graphql-doc-generator-multi-instance.config.js
@@ -1,4 +1,0 @@
-module.exports = [
-  require("./docusaurus2-graphql-doc-generator-multi-instance-1.config.js"),
-  require("./docusaurus2-graphql-doc-generator-multi-instance-2.config.js"),
-];
